chore : domain에서 Entity로 패키지 변경(#1)
