+++ src/main/java/com/study/studyproject/entity/Category.java
@@ -1,4 +1,4 @@
-package com.study.studyproject.domain;
+package com.study.studyproject.entity;
 
 public enum Category{
