+++ src/main/java/com/study/studyproject/entity/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.study.studyproject.domain;
+package com.study.studyproject.entity;
 
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
