+++ src/main/java/com/study/studyproject/entity/reply.java
@@ -1,19 +1,20 @@
-package com.study.studyproject.domain;
+package com.study.studyproject.entity;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 
 @Entity
 @NoArgsConstructor
-public class PostLike extends BaseTimeEntity{
+@Getter
+public class reply extends BaseTimeEntity{
+
     @Id
     @GeneratedValue
     private Long id;
-
+    private String content;
 
 
 }
